refactor verify locales code
